Fix some issues with optee
You have to add "SPD=opteed" to the atf build to have it load optee.

Inexplicably, the tee binary to use with mkimage is tee-raw.bin, not
tee.bin.

optee still doesn't quite work, but it at least starts up and doesn't
crash.  It still won't load an image, though.

Signed-off-by: Corey Minyard <minyard@acm.org>